backoffice homepage and initial navigation stack
