+++ i18n.ts
@@ -15,6 +15,7 @@ export default getRequestConfig(async () => {
     ...(await import(`@/app/(unauthenticated)/reset-password/success/locales/${locale}.json`)).default,
     ...(await import(`@/components/locales/${locale}.json`)).default,
     ...(await import(`@/app/(app)/profile/locales/${locale}.json`)).default,
+    ...(await import(`@/app/(app)/backoffice/locales/${locale}.json`)).default,
   };
 
   return {
